Reduce duplication in booking tests
Ahead of expanding integration tests around booking state changes, we extract and use a helper function to set up the needed stubbed state
