+++ integration_tests/tests/temporary-accommodation/manage/confirmation.cy.ts
@@ -1,12 +1,11 @@
 import Page from '../../../../cypress_shared/pages/page'
 import BookingConfirmationNewPage from '../../../../cypress_shared/pages/temporary-accommodation/manage/bookingConfirmationNew'
 import BookingShowPage from '../../../../cypress_shared/pages/temporary-accommodation/manage/bookingShow'
+import { setupBookingStateStubs } from '../../../../cypress_shared/utils/booking'
 import setupTestUser from '../../../../cypress_shared/utils/setupTestUser'
 import bookingFactory from '../../../../server/testutils/factories/booking'
 import confirmationFactory from '../../../../server/testutils/factories/confirmation'
 import newConfirmationFactory from '../../../../server/testutils/factories/newConfirmation'
-import premisesFactory from '../../../../server/testutils/factories/premises'
-import roomFactory from '../../../../server/testutils/factories/room'
 
 context('Booking confirmation', () => {
   beforeEach(() => {
@@ -19,13 +18,8 @@ context('Booking confirmation', () => {
     cy.signIn()
 
     // And there is a premises, a room, and a provisional booking in the database
-    const premises = premisesFactory.build()
-    const room = roomFactory.build()
     const booking = bookingFactory.provisional().build()
-
-    cy.task('stubSinglePremises', premises)
-    cy.task('stubSingleRoom', { premisesId: premises.id, room })
-    cy.task('stubBooking', { premisesId: premises.id, booking })
+    const { premises, room } = setupBookingStateStubs(booking)
 
     // When I visit the show booking page
     const bookingShow = BookingShowPage.visit(premises, room, booking)
@@ -42,13 +36,8 @@ context('Booking confirmation', () => {
     cy.signIn()
 
     // And there is a premises, a room, and a provisional booking in the database
-    const premises = premisesFactory.build()
-    const room = roomFactory.build()
     const booking = bookingFactory.provisional().build()
-
-    cy.task('stubSinglePremises', premises)
-    cy.task('stubSingleRoom', { premisesId: premises.id, room })
-    cy.task('stubBooking', { premisesId: premises.id, booking })
+    const { premises, room } = setupBookingStateStubs(booking)
 
     // When I visit the booking confirmation page
     const page = BookingConfirmationNewPage.visit(premises, room, booking)
@@ -82,13 +71,8 @@ context('Booking confirmation', () => {
     cy.signIn()
 
     // And there is a premises, a room, and a provisional booking in the database
-    const premises = premisesFactory.build()
-    const room = roomFactory.build()
     const booking = bookingFactory.provisional().build()
-
-    cy.task('stubSinglePremises', premises)
-    cy.task('stubSingleRoom', { premisesId: premises.id, room })
-    cy.task('stubBooking', { premisesId: premises.id, booking })
+    const { premises, room } = setupBookingStateStubs(booking)
 
     // When I visit the booking confirmation page
     const page = BookingConfirmationNewPage.visit(premises, room, booking)
